Update
git-svn-id: https://svn.aros.org/svn/aros/trunk@3957 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/include/defines/mathffp.h
@@ -12,68 +12,68 @@
     Defines
 */
 #define SPAbs(fnum1) \
-    AROS_LC1(LONG, SPAbs, \
-    AROS_LCA(LONG, fnum1, D0), \
+    AROS_LC1(float, SPAbs, \
+    AROS_LCA(float, fnum1, D0), \
     struct MathBase *, MathBase, 9, Mathffp)
 
 #define SPAdd(fnum1, fnum2) \
-    AROS_LC2(LONG, SPAdd, \
-    AROS_LCA(LONG, fnum1, D1), \
-    AROS_LCA(LONG, fnum2, D0), \
+    AROS_LC2(float, SPAdd, \
+    AROS_LCA(float, fnum1, D1), \
+    AROS_LCA(float, fnum2, D0), \
     struct MathBase *, MathBase, 11, Mathffp)
 
 #define SPCeil(y) \
-    AROS_LC1(LONG, SPCeil, \
-    AROS_LCA(LONG, y, D0), \
+    AROS_LC1(float, SPCeil, \
+    AROS_LCA(float, y, D0), \
     struct MathBase *, MathBase, 16, Mathffp)
 
 #define SPCmp(fnum1, fnum2) \
     AROS_LC2(LONG, SPCmp, \
-    AROS_LCA(LONG, fnum1, D0), \
-    AROS_LCA(LONG, fnum2, D1), \
+    AROS_LCA(float, fnum1, D0), \
+    AROS_LCA(float, fnum2, D1), \
     struct MathBase *, MathBase, 7, Mathffp)
 
 #define SPDiv(fnum1, fnum2) \
-    AROS_LC2(LONG, SPDiv, \
-    AROS_LCA(LONG, fnum1, D1), \
-    AROS_LCA(LONG, fnum2, D0), \
+    AROS_LC2(float, SPDiv, \
+    AROS_LCA(float, fnum1, D1), \
+    AROS_LCA(float, fnum2, D0), \
     struct MathBase *, MathBase, 14, Mathffp)
 
 #define SPFix(fnum) \
     AROS_LC1(LONG, SPFix, \
-    AROS_LCA(LONG, fnum, D0), \
+    AROS_LCA(float, fnum, D0), \
     struct MathBase *, MathBase, 5, Mathffp)
 
 #define SPFloor(y) \
-    AROS_LC1(LONG, SPFloor, \
-    AROS_LCA(LONG, y, D0), \
+    AROS_LC1(float, SPFloor, \
+    AROS_LCA(float, y, D0), \
     struct MathBase *, MathBase, 15, Mathffp)
 
 #define SPFlt(inum) \
-    AROS_LC1(LONG, SPFlt, \
+    AROS_LC1(float, SPFlt, \
     AROS_LCA(LONG, inum, D0), \
     struct MathBase *, MathBase, 6, Mathffp)
 
 #define SPMul(fnum1, fnum2) \
-    AROS_LC2(LONG, SPMul, \
-    AROS_LCA(LONG, fnum1, D1), \
-    AROS_LCA(LONG, fnum2, D0), \
+    AROS_LC2(float, SPMul, \
+    AROS_LCA(float, fnum1, D1), \
+    AROS_LCA(float, fnum2, D0), \
     struct MathBase *, MathBase, 13, Mathffp)
 
 #define SPNeg(fnum1) \
-    AROS_LC1(LONG, SPNeg, \
-    AROS_LCA(LONG, fnum1, D0), \
+    AROS_LC1(float, SPNeg, \
+    AROS_LCA(float, fnum1, D0), \
     struct MathBase *, MathBase, 10, Mathffp)
 
 #define SPSub(fnum1, fnum2) \
-    AROS_LC2(LONG, SPSub, \
-    AROS_LCA(LONG, fnum1, D0), \
-    AROS_LCA(LONG, fnum2, D1), \
+    AROS_LC2(float, SPSub, \
+    AROS_LCA(float, fnum1, D0), \
+    AROS_LCA(float, fnum2, D1), \
     struct MathBase *, MathBase, 12, Mathffp)
 
 #define SPTst(fnum) \
     AROS_LC1(LONG, SPTst, \
-    AROS_LCA(LONG, fnum, D1), \
+    AROS_LCA(float, fnum, D1), \
     struct MathBase *, MathBase, 8, Mathffp)
